+++ CPP_Module08/ex01/span.hpp
@@ -6,7 +6,7 @@
 /*   By: sgah <sgah@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/15 19:28:31 by sgah              #+#    #+#             */
-/*   Updated: 2021/06/16 09:49:06 by sgah             ###   ########.fr       */
+/*   Updated: 2021/06/17 11:22:05 by sgah             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ class span {
 		span &	operator=(const span &rhs);
 
 		void	addNumber(int n);
-		int		longestSpan(void);
-		int		shortestSpan(void);
+		long		longestSpan(void);
+		long		shortestSpan(void);
 
 	class TooMuchValueException: public std::exception {
 		public:
